Text changes and removal of authors
I made chnages to the text in general and actually cut out the authors and affiliations. What I am looking for is a general text we can share when we get invited somehwere and I wouldn't put single authors on that. Also, I think OpenNeuroPET should be a single word to not confuse it with OpenNeuro. I also added references, so people can go and find the previous work.