Modified the text
Hi, I went through and did quite some rephrasing to amke the text more readable. Also note, this is not the PET or neuroimaging cloud, so we should be careful with using any tupe of abbreviations. Don't want to end like Gitte that someone comes and talks to us about animals. ;-)